+++ examples/data-sources/sdwan_allow_url_list_policy_object/data-source.tf
@@ -1,0 +1,3 @@
+data "sdwan_allow_url_list_policy_object" "example" {
+  id = "f6b2c44c-693c-4763-b010-895aa3d236bd"
+}