+++ examples/resources/sdwan_allow_url_list_policy_object/resource.tf
@@ -1,0 +1,8 @@
+resource "sdwan_allow_url_list_policy_object" "example" {
+  name = "Example"
+  entries = [
+    {
+      url = "cisco.com"
+    }
+  ]
+}