+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add `sdwan_local_application_list_policy_object` resource and data source
 - Add `sdwan_domain_list_policy_object` resource and data source
 - Add `sdwan_ips_signature_list_policy_object` resource and data source
+- Add `sdwan_allow_url_list_policy_object` resource and data source
 
 ## 0.2.4